* Workround the warnings for clif.c (what to do? the original codes are ugly.. passing values as pointers omgwtfbbq).
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@5856 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/clif.c
@@ -9002,8 +9002,8 @@ void clif_parse_NpcBuyListSend(int fd,struct map_session_data *sd)
 		if((nd = ((struct npc_data *)map_id2bl(sd->npc_shopid))->master_nd)){
 			sprintf(npc_ev, "%s::OnBuyItem", nd->exname);
 			for(i=0;i<n;i++){
-				setd_sub(sd, "@bought_nameid", i, (void *)item_list[i*2+1]);
-				setd_sub(sd, "@bought_quantity", i, (void *)item_list[i*2]);
+				setd_sub(sd, "@bought_nameid", i, (void *)((int)item_list[i*2+1]));
+				setd_sub(sd, "@bought_quantity", i, (void *)((int)item_list[i*2]));
 			}
 			npc_event(sd, npc_ev, 0);
 			fail = 0;
@@ -9040,8 +9040,8 @@ void clif_parse_NpcSellListSend(int fd,struct map_session_data *sd)
 		if((nd = ((struct npc_data *)map_id2bl(sd->npc_shopid))->master_nd)){
 			sprintf(npc_ev, "%s::OnSellItem", nd->exname);
 			for(i=0;i<n;i++){
-				setd_sub(sd, "@sold_nameid", i, (void *)sd->status.inventory[item_list[i*2]-2].nameid);
-				setd_sub(sd, "@sold_quantity", i, (void *)item_list[i*2+1]);
+				setd_sub(sd, "@sold_nameid", i, (void *)((int)sd->status.inventory[item_list[i*2]-2].nameid));
+				setd_sub(sd, "@sold_quantity", i, (void *)((int)item_list[i*2+1]));
 			}
 			npc_event(sd, npc_ev, 0);
 			fail = 0;
